Support MRP arrays with small amounts of invalid submaps (#45)
Invalid submaps are the two submaps for which we do not have submap
UTMs. They affect very little data but we want to still index the
underlying logs, and just leave gaps in UTM-ified MRP wherever the
submap UTM is unavailable.

Also adds altitude to UTM methods for convenience (just the MRP altitude
passed through). Worth noting that the MRP altitude roughly but not
exactly corresponds to altitude above sea level.

Solves #44.

---------

Co-authored-by: Julieta Martinez <jltmtzc@gmail.com>